Begin Code Refactoring, Slight bugs with link to continue figuring out
