+++ LinkDownCommand.cs
@@ -15,7 +15,7 @@ namespace Legend_of_the_Power_Rangers
         }
         public void Execute()
         {
-            this.stateMachine.ChangeState(LinkStateMachine.LinkState.Down);
+            this.stateMachine.ChangeDirection(LinkStateMachine.LinkDirection.Down);
         }
     }
 }
