+++ LinkIdleCommand.cs
@@ -15,7 +15,10 @@ namespace Legend_of_the_Power_Rangers
         }
         public void Execute()
         {
-            this.stateMachine.ChangeState(LinkStateMachine.LinkState.Idle);
+            this.stateMachine.ChangeDirection(LinkStateMachine.LinkDirection.Idle);
+            this.stateMachine.ChangeAction(LinkStateMachine.LinkAction.Idle);
+
+
         }
     }
 }
