+++ LinkSpritesClasses/LinkDownSprite.cs
