+++ LinkSpritesClasses/LinkLeftSprite.cs
