+++ LinkSpritesClasses/LinkRightSprite.cs
